Simplify the implementation of the two transformers which work with stop words, by providing the common code in a template pattern.
