+++ silk-rules/src/main/scala/org/silkframework/rule/plugins/RulePlugins.scala
@@ -80,7 +80,7 @@ class RulePlugins extends PluginModule {
         classOf[RemoveSpecialCharsTransformer] ::
         classOf[ConvertCharsetTransformer] ::
         classOf[RemoveValues] ::
-        classOf[RemoveRemoteStopwordsTransformer] ::
+        classOf[RemoveRemoteStopWordsTransformer] ::
         classOf[RemoveEmptyValues] ::
         classOf[RemoveParentheses] ::
         classOf[TrimTransformer] ::
