+++ silk-workspace/src/main/scala/org/silkframework/workspace/WorkspacePlugins.scala
@@ -1,7 +1,7 @@
 package org.silkframework.workspace
 
 import org.silkframework.plugins.dataset.DatasetTypeAutoCompletionProvider
-import org.silkframework.plugins.filter.RemoveStopwordsTransformer
+import org.silkframework.plugins.filter.RemoveStopWordsTransformer
 import org.silkframework.plugins.transformer.value.ReadParameter
 import org.silkframework.runtime.plugin.{AnyPlugin, PluginModule}
 import org.silkframework.workspace.activity.dataset.Types.TypesFormat
@@ -75,7 +75,7 @@ class WorkspacePlugins extends PluginModule {
 
   def rulePlugins: List[Class[_ <: AnyPlugin]] = {
     classOf[ReadParameter] ::
-    classOf[RemoveStopwordsTransformer] ::
+    classOf[RemoveStopWordsTransformer] ::
     Nil
   }
